sql_chat_agent/requirements.txt
streamlit
sqlalchemy
python-dotenv
langchain
langchain-community
sentence-transformers
faiss-cpu
openai